+++ src/main/java/com/tienda/inventario/repositories/VentaClienteRepository.java
@@ -8,7 +8,7 @@ import com.tienda.inventario.entities.VentaCliente;
 import com.tienda.inventario.entities.Venta;
 import com.tienda.inventario.entities.CuentaCliente;
 
-public interface VentaClienteRepository extends JpaRepository<VentaCliente, Long> {
+public interface VentaClienteRepository extends JpaRepository<VentaCliente, Integer> {
 
     List<VentaCliente> findByVenta(Venta venta);
